refactor(wallet): extract Client::TransactionOps concern module (#577)
Move the 7 BRC-100 transaction public methods and all their private
helpers (~900 lines) out of client.rb into a new include module,
reducing client.rb from 1416 to 470 lines.
